pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]

requires-python = ">=3.11.3"

name = "xy_django_app_api_auth"

authors = [{ name = "helios", email = "yuyang.0515@qq.com" }]
description = "后台账户数据模型"

dynamic = ["version"]

classifiers = ["Programming Language :: Python :: 3"]

dependencies = [
    "django",
    "djangorestframework",
    "xy_type",
    "xy_string",
    "xy_django_model",
    "xy_admin_model_serializer",
]

readme = "README.md"

[project.scripts]

[tool.setuptools.packages.find]
where = ["./"]
include = ["xy_django_app_api_auth"] # alternatively: `exclude = ["additional*"]`
namespaces = false

[tool.setuptools]
include-package-data = true

[tool.setuptools.dynamic]
version = { file = "VERSION" }
